add initial test and fix file generation
