+++ cli/__tests__/example.yaml
@@ -1,0 +1,79 @@
+name: Docs
+
+# why isn't this just on: push: ?
+# because we want to test different build patterns, that's why
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - 'docs/**'
+      - '.stoat/**'
+      - '.github/workflows/docs.yaml'
+  pull_request:
+    paths:
+      - 'docs/**'
+      - '.stoat/**'
+      - '.github/workflows/docs.yaml'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      # misdirection
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Set up Node
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18
+
+      - name: Setup cache
+        uses: actions/cache@v3
+        with:
+          path: |
+            ${{ github.workspace }}/docs/node_modules
+          key: ${{ runner.os }}-node-16-${{ hashFiles('docs/package-lock.json') }}
+
+      - name: Build
+        run: |
+          cd docs
+          npm install
+          npm run build
+  build2:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+  build3:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Set up Node
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18
+
+      - name: Setup cache
+        uses: actions/cache@v3
+        with:
+          path: |
+            ${{ github.workspace }}/docs/node_modules
+          key: ${{ runner.os }}-node-16-${{ hashFiles('docs/package-lock.json') }}
+
+      - name: Build
+        run: |
+          cd docs
+          npm install
+          npm run build
+
+      - name: Run Stoat Action
+        uses: stoat-dev/stoat-action@v0
+        if: always()
